+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Family Robot styles from [@JonDotsoy][]
 - Z-Depth styles from [@JonDotsoy][]
 - reset.css style from [@JonDotsoy][]
-- Colors Material Design styles [@JonDotsoy][]
+- Colors Material Design style from [@JonDotsoy][]
+- `equip-image()` postcss plugin from [@JonDotsoy][]
 
 [@JonDotsoy]: https://github.com/jondotsoy